Process Approval Pending Status Fix
This is a fix for an issue whereby the process_approval_statuses status column does not get updated to pending.
When an approval process is no longer in the submitted status and the isApprovalCompleted method is false the status should be pending,
unfortunately due to a type mismatch between the enum casted status column and the ApprovalStatusEnum the condition is never true.
Because of this the process_approval_statuses only moves between Submitted and Approved.

This fixes that issue by ensuring that they are of the same type for proper comparsiom.
The fix now introduces a new issue where the status never gets to Apporved, this is because we check the isApprovalCompleted method before we update the status.
This leads to the final approval status always been set to Pending because it assumes that the process is not completed yet.
To fix this I have introduced a nullable variable for the isApprovalCompleted method that can take the current steps and use it to confirm if the process is completed.